+++ app/Http/Controllers/Admin/AdminJobGenderController.php
@@ -59,12 +59,12 @@ class AdminJobGenderController extends Controller
 
     public function delete($id)
     {
-        // $check = Job::where('job_gender_id', $id)->count();
-        // if ($check > 0) {
-        //     return redirect()
-        //         ->back()
-        //         ->with('error', 'You can not delete this item, because this is used in another place.');
-        // }
+        $check = Job::where('job_gender_id', $id)->count();
+        if ($check > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'You can not delete this item, because this is used in another place.');
+        }
 
         JobGender::where('id', $id)->delete();
         return redirect()
