+++ app/Http/Controllers/Admin/AdminJobLocationController.php
@@ -60,12 +60,12 @@ class AdminJobLocationController extends Controller
 
     public function delete($id)
     {
-        // $check = Job::where('job_location_id', $id)->count();
-        // if ($check > 0) {
-        //     return redirect()
-        //         ->back()
-        //         ->with('error', 'You can not delete this item, because this is used in another place.');
-        // }
+        $check = Job::where('job_location_id', $id)->count();
+        if ($check > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'You can not delete this item, because this is used in another place.');
+        }
 
         JobLocation::where('id', $id)->delete();
         return redirect()
