+++ app/Http/Controllers/Admin/AdminJobSalaryRangeController.php
@@ -58,12 +58,12 @@ class AdminJobSalaryRangeController extends Controller
 
     public function delete($id)
     {
-        // $check = Job::where('job_salary_range_id', $id)->count();
-        // if ($check > 0) {
-        //     return redirect()
-        //         ->back()
-        //         ->with('error', 'You can not delete this item, because this is used in another place.');
-        // }
+        $check = Job::where('job_salary_range_id', $id)->count();
+        if ($check > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'You can not delete this item, because this is used in another place.');
+        }
 
         JobSalaryRange::where('id', $id)->delete();
         return redirect()
